+++ src/setuptools_scm/_overrides.py
@@ -1,12 +1,15 @@
 from __future__ import annotations
 
 import os
+import re
 from typing import Any
 
 from . import _config
+from . import _log
 from . import version
 from ._integration.pyproject_reading import lazy_toml_load
-from ._trace import trace
+
+log = _log.log.getChild("overrides")
 
 PRETEND_KEY = "SETUPTOOLS_SCM_PRETEND_VERSION"
 PRETEND_KEY_NAMED = PRETEND_KEY + "_FOR_{name}"
@@ -16,7 +19,10 @@ def read_named_env(
     *, tool: str = "SETUPTOOLS_SCM", name: str, dist_name: str | None
 ) -> str | None:
     if dist_name is not None:
-        val = os.environ.get(f"{tool}_{name}_FOR_{dist_name.upper()}")
+        # Normalize the dist name as per PEP 503.
+        normalized_dist_name = re.sub(r"[-_.]+", "-", dist_name)
+        env_var_dist_name = normalized_dist_name.replace("-", "_").upper()
+        val = os.environ.get(f"{tool}_{name}_FOR_{env_var_dist_name}")
         if val is not None:
             return val
     return os.environ.get(f"{tool}_{name}")
@@ -30,7 +36,7 @@ def _read_pretended_version_for(
     tries ``SETUPTOOLS_SCM_PRETEND_VERSION``
     and ``SETUPTOOLS_SCM_PRETEND_VERSION_FOR_$UPPERCASE_DIST_NAME``
     """
-    trace("dist name:", config.dist_name)
+    log.debug("dist name: %s", config.dist_name)
 
     pretended = read_named_env(name="PRETEND_VERSION", dist_name=config.dist_name)