+++ src/setuptools_scm/_version_cls.py
@@ -38,8 +38,9 @@ def _version_as_tuple(version_str: str) -> tuple[int | str, ...]:
     try:
         parsed_version = Version(version_str)
     except InvalidVersion:
-        log = getLogger("setuptools_scm")
-        log.exception("failed to parse version %s", version_str)
+        log = getLogger(__name__).parent
+        assert log is not None
+        log.error("failed to parse version %s", version_str)
         return (version_str,)
     else:
         version_fields: tuple[int | str, ...] = parsed_version.release