+++ src/setuptools_scm/hg_git.py
@@ -1,19 +1,25 @@
 from __future__ import annotations
 
+import logging
 import os
 from contextlib import suppress
 from datetime import date
-from datetime import datetime
+from pathlib import Path
 
 from . import _types as _t
-from ._trace import trace
+from ._run_cmd import CompletedProcess as _CompletedProcess
+from ._run_cmd import require_command
+from ._run_cmd import run as _run
 from .git import GitWorkdir
 from .hg import HgWorkdir
-from .utils import _CmdResult
-from .utils import do_ex
-from .utils import require_command
 
-_FAKE_GIT_DESCRIBE_ERROR = _CmdResult("<>hg git failed", "", 1)
+log = logging.getLogger(__name__)
+
+_FAKE_GIT_DESCRIBE_ERROR = _CompletedProcess(
+    "fake git describe output for hg",
+    1,
+    "<>hg git failed to describe",
+)
 
 
 class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
@@ -21,29 +27,27 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
 
     @classmethod
     def from_potential_worktree(cls, wd: _t.PathT) -> GitWorkdirHgClient | None:
-        require_command(cls.COMMAND)
-        root, _, ret = do_ex(["hg", "root"], wd)
-        if ret:
+        require_command("hg")
+        res = _run(["hg", "root"], cwd=wd).parse_success(parse=Path)
+        if res is None:
             return None
-        return cls(root)
+        return cls(res)
 
     def is_dirty(self) -> bool:
-        out, _, _ = self.do_ex('hg id -T "{dirty}"')
-        return bool(out)
+        res = _run(["hg", "id", "-T", "{dirty}"], cwd=self.path, check=True)
+        return bool(res.stdout)
 
     def get_branch(self) -> str | None:
-        res = self.do_ex('hg id -T "{bookmarks}"')
+        res = _run(["hg", "id", "-T", "{bookmarks}"], cwd=self.path)
         if res.returncode:
-            trace("branch err", res)
+            log.info("branch err %s", res)
             return None
-        return res.out
+        return res.stdout
 
     def get_head_date(self) -> date | None:
-        date_part, err, ret = self.do_ex('hg log -r . -T "{shortdate(date)}"')
-        if ret:
-            trace("head date err", date_part, err, ret)
-            return None
-        return datetime.strptime(date_part, r"%Y-%m-%d").date()
+        return _run('hg log -r . -T "{shortdate(date)}"', cwd=self.path).parse_success(
+            parse=date.fromisoformat, error_msg="head date err"
+        )
 
     def is_shallow(self) -> bool:
         return False
@@ -52,11 +56,11 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
         pass
 
     def get_hg_node(self) -> str | None:
-        node, _, ret = self.do_ex('hg log -r . -T "{node}"')
-        if not ret:
-            return node
-        else:
+        res = _run('hg log -r . -T "{node}"', cwd=self.path)
+        if res.returncode:
             return None
+        else:
+            return res.stdout
 
     def _hg2git(self, hg_node: str) -> str | None:
         with suppress(FileNotFoundError):
@@ -76,11 +80,11 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
 
         if git_node is None:
             # trying again after hg -> git
-            self.do_ex("hg gexport")
+            _run(["hg", "gexport"], cwd=self.path)
             git_node = self._hg2git(hg_node)
 
             if git_node is None:
-                trace("Cannot get git node so we use hg node", hg_node)
+                log.debug("Cannot get git node so we use hg node %s", hg_node)
 
                 if hg_node == "0" * len(hg_node):
                     # mimic Git behavior
@@ -91,17 +95,17 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
         return git_node[:7]
 
     def count_all_nodes(self) -> int:
-        revs, _, _ = self.do_ex(["hg", "log", "-r", "ancestors(.)", "-T", "."])
-        return len(revs)
+        res = _run(["hg", "log", "-r", "ancestors(.)", "-T", "."], cwd=self.path)
+        return len(res.stdout)
 
-    def default_describe(self) -> _CmdResult:
+    def default_describe(self) -> _CompletedProcess:
         """
         Tentative to reproduce the output of
 
         `git describe --dirty --tags --long --match *[0-9]*`
 
         """
-        hg_tags_str, _, ret = self.do_ex(
+        res = _run(
             [
                 "hg",
                 "log",
@@ -109,16 +113,17 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
                 "(reverse(ancestors(.)) and tag(r're:v?[0-9].*'))",
                 "-T",
                 "{tags}{if(tags, ' ', '')}",
-            ]
+            ],
+            cwd=self.path,
         )
-        if ret:
+        if res.returncode:
             return _FAKE_GIT_DESCRIBE_ERROR
-        hg_tags: list[str] = hg_tags_str.split()
+        hg_tags: list[str] = res.stdout.split()
 
         if not hg_tags:
             return _FAKE_GIT_DESCRIBE_ERROR
 
-        with open(os.path.join(self.path, ".hg/git-tags")) as fp:
+        with self.path.joinpath(".hg/git-tags").open() as fp:
             git_tags: dict[str, str] = dict(line.split()[::-1] for line in fp)
 
         tag: str
@@ -127,13 +132,13 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
                 tag = hg_tag
                 break
         else:
-            trace("tag not found", hg_tags, git_tags)
+            logging.warning("tag not found hg=%s git=%s", hg_tags, git_tags)
             return _FAKE_GIT_DESCRIBE_ERROR
 
-        out, _, ret = self.do_ex(["hg", "log", "-r", f"'{tag}'::.", "-T", "."])
-        if ret:
+        res = _run(["hg", "log", "-r", f"'{tag}'::.", "-T", "."], cwd=self.path)
+        if res.returncode:
             return _FAKE_GIT_DESCRIBE_ERROR
-        distance = len(out) - 1
+        distance = len(res.stdout) - 1
 
         node = self.node()
         assert node is not None
@@ -141,5 +146,10 @@ class GitWorkdirHgClient(GitWorkdir, HgWorkdir):
 
         if self.is_dirty():
             desc += "-dirty"
-        trace("desc", desc)
-        return _CmdResult(desc, "", 0)
+        log.debug("faked describe %r", desc)
+        return _CompletedProcess(
+            ["setuptools-scm", "faked", "describe"],
+            returncode=0,
+            stdout=desc,
+            stderr="",
+        )