+++ testing/test_regressions.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import os
 import pprint
 import subprocess
 import sys
@@ -16,7 +15,6 @@ from pathlib import Path
 import pytest
 
 from setuptools_scm import Configuration
-from setuptools_scm import get_version
 from setuptools_scm.git import parse
 from setuptools_scm._run_cmd import run
 
@@ -48,28 +46,7 @@ def test_pkginfo_noscmroot(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> N
         pass
     else:
         res = run([sys.executable, "setup.py", "--version"], p)
-        assert res.stdout == "0.1.dev0"
-
-
-def test_pip_egg_info(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    """if we are indeed a sdist, the root does not apply"""
-
-    # we should get the version from pkg-info if git is broken
-    p = tmp_path.joinpath("sub/package")
-    p.mkdir(parents=True)
-    tmp_path.joinpath(".git").mkdir()
-    p.joinpath("setup.py").write_text(
-        "from setuptools import setup;" 'setup(use_scm_version={"root": ".."})'
-    )
-
-    with pytest.raises(LookupError):
-        get_version(root=os.fspath(p), fallback_root=os.fspath(p))
-
-    bad_egg_info = p.joinpath("pip-egg-info/random.egg-info/")
-    bad_egg_info.mkdir(parents=True)
-
-    bad_egg_info.joinpath("PKG-INFO").write_text("Version: 1.0")
-    assert get_version(root=os.fspath(p), fallback_root=os.fspath(p)) == "1.0"
+        assert res.stdout == "0.1.dev0+d20090213"
 
 
 @pytest.mark.issue(164)
@@ -113,7 +90,7 @@ def test_case_mismatch_on_windows_git(tmp_path: Path) -> None:
 
 
 def test_entrypoints_load() -> None:
-    d = distribution("setuptools-scm")  # type: ignore [no-untyped-call]
+    d = distribution("setuptools-scm")
 
     eps = d.entry_points
     failed: list[tuple[EntryPoint, Exception]] = []